add test changes to remove areas outside workspace
May have fixed out of bounds error for arccos, but ideally only want to
calculate this heatmap for values that are inside the workspace to lower
perf cost and avoid any other errors. Need to get update texture working
so the workspace can be plot just as solid colors to test its working.